+++ Hippo.Email/Views/Emails/Sample_mjml.cshtml
@@ -1,4 +1,4 @@
-﻿@using Hippo.Email.Views.Shared
+﻿@using Hippo.Email.Models
 @model Hippo.Email.Models.SampleModel
 
 @{
