+++ Hippo.Email/Views/Shared/_EmailInnerDetails_mjml.cshtml
@@ -1,4 +1,4 @@
-﻿@using  Hippo.Email.Views.Shared
+﻿@using  Hippo.Email.Models
 @model EmailInnerDetailsModel
 
 <mj-text padding-top="0px" padding-left="0px" padding-bottom="5px">
